+++ benches/macros.rs
@@ -12,6 +12,12 @@ use std::hint::black_box;
 
 use shwild::shwild_matches;
 
+#[cfg(feature = "assertions")]
+use shwild::{
+    assert_shwild_matches,
+    assert_shwild_not_matches,
+};
+
 
 mod constants {
     #![allow(non_upper_case_globals)]
@@ -307,6 +313,374 @@ mod macros_benches {
 }
 
 
+
+#[cfg(feature = "assertions")]
+mod assertion_benches {
+    #![allow(non_snake_case)]
+
+    use super::*;
+
+
+    pub fn assert_shwild_matches_input_empty(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = "";
+
+        c.bench_function("`assert_shwild_matches!()` - empty string", |b| {
+            b.iter(|| {
+                assert_shwild_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_input_literal_small(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = "";
+
+        c.bench_function("`assert_shwild_matches!()` - literal (small)", |b| {
+            b.iter(|| {
+                assert_shwild_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_input_literal_medium(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = "";
+
+        c.bench_function("`assert_shwild_matches!()` - literal (medium)", |b| {
+            b.iter(|| {
+                assert_shwild_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_input_literal_large(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = "";
+
+        c.bench_function("`assert_shwild_matches!()` - literal (large)", |b| {
+            b.iter(|| {
+                assert_shwild_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_against_nrange_continuum_simple(c : &mut Criterion) {
+        let pattern = constants::patterns::NRANGE_CONTINUUM_SIMPLE;
+
+        let inputs = [
+            // insert list:
+            " ",
+            "_",
+        ];
+
+        c.bench_function("`assert_shwild_matches!()` - nrange continuum (simple)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_against_nrange_continuum_reverse(c : &mut Criterion) {
+        let pattern = constants::patterns::NRANGE_CONTINUUM_REVERSE;
+
+        let inputs = [
+            // insert list:
+            " ",
+            "_",
+        ];
+
+        c.bench_function("`assert_shwild_matches!()` - nrange continuum (reverse)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_against_nrange_continuum_crosscase(c : &mut Criterion) {
+        let pattern = constants::patterns::NRANGE_CONTINUUM_CROSSCASE;
+
+        let inputs = [
+            // insert list:
+            " ",
+            "_",
+        ];
+
+        c.bench_function("`assert_shwild_matches!()` - nrange continuum (crosscase)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_against_range_continuum_simple(c : &mut Criterion) {
+        let pattern = constants::patterns::RANGE_CONTINUUM_SIMPLE;
+
+        let inputs = [
+            // insert list:
+            "a",
+            "b",
+            "c",
+            "d",
+        ];
+
+        c.bench_function("`assert_shwild_matches!()` - range continuum (simple)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_against_range_continuum_reverse(c : &mut Criterion) {
+        let pattern = constants::patterns::RANGE_CONTINUUM_REVERSE;
+
+        let inputs = [
+            // insert list:
+            "a",
+            "b",
+            "c",
+            "d",
+        ];
+
+        c.bench_function("`assert_shwild_matches!()` - range continuum (reverse)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_against_range_continuum_crosscase(c : &mut Criterion) {
+        let pattern = constants::patterns::RANGE_CONTINUUM_CROSSCASE;
+
+        let inputs = [
+            // insert list:
+            "a",
+            "b",
+            "c",
+            "d",
+            "A",
+            "B",
+            "C",
+            "D",
+        ];
+
+        c.bench_function("`assert_shwild_matches!()` - range continuum (crosscase)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_matches_test_against_pattern_WindowsPath(c : &mut Criterion) {
+        let pattern = constants::patterns::WINDOWS_PATH;
+        let inputs = constants::windows_path_inputs::MATCHING;
+
+        c.bench_function("`assert_shwild_matches!()` - Windows Path", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_input_empty(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = " ";
+
+        c.bench_function("`assert_shwild_not_matches!()` - empty string", |b| {
+            b.iter(|| {
+                assert_shwild_not_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_input_literal_small(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = " ";
+
+        c.bench_function("`assert_shwild_not_matches!()` - literal (small)", |b| {
+            b.iter(|| {
+                assert_shwild_not_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_input_literal_medium(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = " ";
+
+        c.bench_function("`assert_shwild_not_matches!()` - literal (medium)", |b| {
+            b.iter(|| {
+                assert_shwild_not_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_input_literal_large(c : &mut Criterion) {
+        let pattern = constants::EMPTY_STRING;
+        let input = " ";
+
+        c.bench_function("`assert_shwild_not_matches!()` - literal (large)", |b| {
+            b.iter(|| {
+                assert_shwild_not_matches!(black_box(pattern), black_box(input));
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_against_nrange_continuum_simple(c : &mut Criterion) {
+        let pattern = constants::patterns::NRANGE_CONTINUUM_SIMPLE;
+
+        let inputs = [
+            // insert list:
+            "a",
+            "b",
+            "c",
+            "d",
+            "",
+            "aa",
+        ];
+
+        c.bench_function("`assert_shwild_not_matches!()` - nrange continuum (simple)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_not_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_against_nrange_continuum_reverse(c : &mut Criterion) {
+        let pattern = constants::patterns::NRANGE_CONTINUUM_REVERSE;
+
+        let inputs = [
+            // insert list:
+            "a",
+            "b",
+            "c",
+            "d",
+            "",
+            "aa",
+        ];
+
+        c.bench_function("`assert_shwild_not_matches!()` - nrange continuum (reverse)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_not_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_against_nrange_continuum_crosscase(c : &mut Criterion) {
+        let pattern = constants::patterns::NRANGE_CONTINUUM_CROSSCASE;
+
+        let inputs = [
+            // insert list:
+            "a",
+            "b",
+            "c",
+            "d",
+            "",
+            "aa",
+        ];
+
+        c.bench_function("`assert_shwild_not_matches!()` - nrange continuum (crosscase)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_not_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_against_range_continuum_simple(c : &mut Criterion) {
+        let pattern = constants::patterns::RANGE_CONTINUUM_SIMPLE;
+
+        let inputs = [
+            // insert list:
+            "",
+            " ",
+            "aa",
+            "_",
+        ];
+
+        c.bench_function("`assert_shwild_not_matches!()` - range continuum (simple)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_not_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_against_range_continuum_reverse(c : &mut Criterion) {
+        let pattern = constants::patterns::RANGE_CONTINUUM_REVERSE;
+
+        let inputs = [
+            // insert list:
+            "",
+            " ",
+            "aa",
+            "_",
+        ];
+
+        c.bench_function("`assert_shwild_not_matches!()` - range continuum (reverse)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_not_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_against_range_continuum_crosscase(c : &mut Criterion) {
+        let pattern = constants::patterns::RANGE_CONTINUUM_CROSSCASE;
+
+        let inputs = [
+            // insert list:
+            "",
+            " ",
+            "aa",
+            "_",
+        ];
+
+        c.bench_function("`assert_shwild_not_matches!()` - range continuum (crosscase)", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_not_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+
+    pub fn assert_shwild_not_matches_test_against_pattern_WindowsPath(c : &mut Criterion) {
+        let pattern = constants::patterns::WINDOWS_PATH;
+        let inputs = constants::windows_path_inputs::NOT_MATCHING;
+
+        c.bench_function("`assert_shwild_not_matches!()` - Windows Path", |b| {
+            b.iter(|| {
+                for input in &inputs {
+                    assert_shwild_not_matches!(black_box(pattern), black_box(input));
+                }
+            })
+        });
+    }
+}
+
+
 criterion_group!(
     shwild_matches_benches,
     macros_benches::shwild_matches_input_empty,
@@ -322,4 +696,44 @@ criterion_group!(
     macros_benches::shwild_matches_against_range_continuum_crosscase,
 );
 
+#[cfg(feature = "assertions")]
+criterion_group!(
+    assert_shwild_matches_benches,
+    assertion_benches::assert_shwild_matches_input_empty,
+    assertion_benches::assert_shwild_matches_input_literal_small,
+    assertion_benches::assert_shwild_matches_input_literal_medium,
+    assertion_benches::assert_shwild_matches_input_literal_large,
+    assertion_benches::assert_shwild_matches_test_against_pattern_WindowsPath,
+    assertion_benches::assert_shwild_matches_against_nrange_continuum_simple,
+    assertion_benches::assert_shwild_matches_against_nrange_continuum_reverse,
+    assertion_benches::assert_shwild_matches_against_nrange_continuum_crosscase,
+    assertion_benches::assert_shwild_matches_against_range_continuum_simple,
+    assertion_benches::assert_shwild_matches_against_range_continuum_reverse,
+    assertion_benches::assert_shwild_matches_against_range_continuum_crosscase,
+);
+
+#[cfg(feature = "assertions")]
+criterion_group!(
+    assert_shwild_not_matches_benches,
+    assertion_benches::assert_shwild_not_matches_input_empty,
+    assertion_benches::assert_shwild_not_matches_input_literal_small,
+    assertion_benches::assert_shwild_not_matches_input_literal_medium,
+    assertion_benches::assert_shwild_not_matches_input_literal_large,
+    assertion_benches::assert_shwild_not_matches_test_against_pattern_WindowsPath,
+    assertion_benches::assert_shwild_not_matches_against_nrange_continuum_simple,
+    assertion_benches::assert_shwild_not_matches_against_nrange_continuum_reverse,
+    assertion_benches::assert_shwild_not_matches_against_nrange_continuum_crosscase,
+    assertion_benches::assert_shwild_not_matches_against_range_continuum_simple,
+    assertion_benches::assert_shwild_not_matches_against_range_continuum_reverse,
+    assertion_benches::assert_shwild_not_matches_against_range_continuum_crosscase,
+);
+
+#[cfg(feature = "assertions")]
+criterion_main!(
+    shwild_matches_benches,
+    assert_shwild_matches_benches,
+    assert_shwild_not_matches_benches,
+);
+
+#[cfg(not(feature = "assertions"))]
 criterion_main!(shwild_matches_benches);